+++ lib/init.ts
@@ -1,8 +1,5 @@
-import { SearchClient } from 'algoliasearch'
-
 import { isUndefined, isString, isNumber } from "./utils";
 import { DEFAULT_ALGOLIA_AGENT } from "./_algoliaAgent";
-import {bindSearchClient } from './_bindSearchClient'
 
 type InsightRegion = "de" | "us";
 const SUPPORTED_REGIONS: InsightRegion[] = ["de", "us"];
@@ -16,7 +13,6 @@ export interface InitParams {
   cookieDuration?: number;
   region?: InsightRegion;
   userToken?: string;
-  searchClient?: SearchClient;
 }
 
 /**
@@ -87,9 +83,5 @@ You can visit https://algolia.com/events/debugger instead.`);
     this.setUserToken(options.userToken);
   } else if (!this._userToken && !this._userHasOptedOut && this._useCookie) {
     this.setAnonymousUserToken();
-  }
-
-  if (options.searchClient) {
-    bindSearchClient.call(this, options.searchClient)
   }
 }
